+++ src/main/java/org/sopt/kareer/domain/roadmap/service/PhaseActionQueryService.java
@@ -1,57 +1,37 @@
 package org.sopt.kareer.domain.roadmap.service;
 
-import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.sopt.kareer.domain.roadmap.dto.response.AiGuideResponse;
-import org.sopt.kareer.domain.roadmap.entity.ActionItem;
 import org.sopt.kareer.domain.roadmap.entity.PhaseAction;
 import org.sopt.kareer.domain.roadmap.entity.PhaseActionTranslation;
 import org.sopt.kareer.domain.roadmap.entity.enums.RoadmapActiveStatus;
 import org.sopt.kareer.domain.roadmap.exception.RoadMapException;
 import org.sopt.kareer.domain.roadmap.exception.RoadmapErrorCode;
-import org.sopt.kareer.domain.roadmap.repository.ActionItemRepository;
 import org.sopt.kareer.domain.roadmap.repository.PhaseActionGuidelineRepository;
 import org.sopt.kareer.domain.roadmap.repository.PhaseActionGuidelineTranslationRepository;
 import org.sopt.kareer.domain.roadmap.repository.PhaseActionMistakeRepository;
 import org.sopt.kareer.domain.roadmap.repository.PhaseActionMistakeTranslationRepository;
 import org.sopt.kareer.domain.roadmap.repository.PhaseActionRepository;
 import org.sopt.kareer.domain.roadmap.repository.PhaseActionTranslationRepository;
+import org.sopt.kareer.domain.roadmap.service.dto.response.AiGuideData;
 import org.springframework.context.i18n.LocaleContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.List;
+
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
-public class PhaseActionService {
+public class PhaseActionQueryService {
 
     private final PhaseActionRepository phaseActionRepository;
-    private final ActionItemRepository actionItemRepository;
-    private final PhaseActionMistakeRepository mistakeRepository;
     private final PhaseActionGuidelineRepository guidelineRepository;
+    private final PhaseActionMistakeRepository mistakeRepository;
     private final PhaseActionTranslationRepository phaseActionTranslationRepository;
     private final PhaseActionGuidelineTranslationRepository guidelineTranslationRepository;
     private final PhaseActionMistakeTranslationRepository mistakeTranslationRepository;
 
-    @Transactional
-    public void createPhaseActionTodo(Long memberId, Long phaseActionId) {
-        PhaseAction phaseAction = phaseActionRepository.findByIdAndMemberIdAndRoadmapStatus(phaseActionId, memberId, RoadmapActiveStatus.ACTIVE)
-                .orElseThrow(() -> new RoadMapException(RoadmapErrorCode.PHASE_ACTION_NOT_FOUND));
-
-        if (phaseAction.getAdded()) {
-            throw new RoadMapException(RoadmapErrorCode.TODO_ALREADY_ADDED);
-        }
-
-        List<ActionItem> actionItems = actionItemRepository.findAllByPhaseActionId(phaseActionId);
-
-        for (ActionItem actionItem : actionItems) {
-            actionItem.activate();
-        }
-
-        phaseAction.markAdded();
-    }
-
-    public AiGuideResponse getAiGuide(Long memberId, Long phaseActionId) {
+    public AiGuideData getAiGuide(Long memberId, Long phaseActionId) {
         PhaseAction phaseAction = phaseActionRepository.findByIdAndMemberIdAndRoadmapStatus(phaseActionId, memberId, RoadmapActiveStatus.ACTIVE)
                 .orElseThrow(() -> new RoadMapException(RoadmapErrorCode.PHASE_ACTION_NOT_FOUND));
 
@@ -73,6 +53,6 @@ public class PhaseActionService {
                 mistakeTranslationRepository.findContentByPhaseActionIdAndLanguage(phaseActionId, language);
         if (!translatedMistakes.isEmpty()) mistakes = translatedMistakes;
 
-        return new AiGuideResponse(importance, mistakes, guidelines);
+        return new AiGuideData(importance, guidelines, mistakes);
     }
 }
